requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
pydantic
openai
google-generativeai